Exit successfully on init container errors To avoid stopping customer applications from starting if there are issues.

+++ pkg/controller/namespace/namespace_controller_test.go
@@ -105,10 +105,16 @@ if [[ "${proxy}" != "" ]]; then
 fi
 
 echo "Downloading OneAgent package..."
-curl "${curl_params[@]}"
+if ! curl "${curl_params[@]}"; then
+	echo "Failed to download the OneAgent package."
+	exit 0
+fi
 
 echo "Unpacking OneAgent package..."
-unzip -o -d "${target_dir}" "${archive}"
+if ! unzip -o -d "${target_dir}" "${archive}"; then
+	echo "Failed to unpack the OneAgent package."
+	exit 0
+fi
 rm -f "${archive}"
 
 echo "Configuring OneAgent..."
